+++ bg/battlemetrics/getMyServers.js
@@ -1,6 +1,10 @@
 export async function getMyServers(BMToken) {
-    const response = await fetch(`https://api.battlemetrics.com/servers?filter[rcon]=true&page[size]=100&access_token=${BMToken}`);
-    if (!response.ok) return;
-
-    return await response.json();
+    const params = new URLSearchParams({
+        "filter[rcon]": "true",
+        "page[size]": "100",
+        "access_token": BMToken,
+    });
+    const response = await fetch(`https://api.battlemetrics.com/servers?${params}`);
+    if (!response.ok) return undefined;
+    return response.json();
 }