+++ bg/bg.js
@@ -23,7 +23,6 @@ chrome.runtime.onMessage.addListener(async (req, sender) => {
     if (req.type.startsWith("BME_STEAM_FRIENDLIST")) return sendFriendlistFromSteam(req.subject, req.apiKey, sender, returnObject);
     // if (req.type.startsWith("BME_RUST_API_FRIENDLIST")) return sendFriendlistFromRustApi(req.subject, req.apiKey, sender, returnObject)
     // if (req.type.startsWith("BME_RUST_API_AVATARS")) return sendAvatarsFromRustApi(req.subject, req.apiKey, sender, returnObject)
-    if (req.type.startsWith("BME_PREMIUM_STATUS")) return sendPremiumStatus(req.subject, sender, returnObject)
     if (req.type.startsWith("BME_PROXYCHECK")) return sendProxyCheck(req.subject, req.apiKey, sender, returnObject)
     if (req.type.startsWith("BME_PLAYER_SUMMARIES")) return sendSteamPlayerSummaries(req.subject, req.apiKey, sender, returnObject);
     if (req.type.startsWith("BME_BAN_SUMMARIES")) return sendSteamPlayerBanSummaries(req.subject, req.apiKey, sender, returnObject);
@@ -73,48 +72,7 @@ async function sendFriendlistFromSteam(steamId, apiKey, sender, returnObject) {
         return chrome.tabs.sendMessage(sender.tab.id, returnObject);
     }
 }
-// async function sendFriendlistFromRustApi(steamId, apiKey, sender, returnObject) {
-//     try {
-//         const resp = await fetch(``);
-//         if (resp?.status !== 200) throw new Error(`Requesting Rust Api Friends Failed | steamId: ${steamId} | API KEY: ${apiKey.substring(0, 10)}... | Status: ${resp?.status}`)
 
-//         const data = await resp.json();
-//         returnObject.status = "OK";
-//         returnObject.value = data.data.friends;
-//         return chrome.tabs.sendMessage(sender.tab.id, returnObject);
-//     } catch (error) {
-//         console.error(error);
-//         returnObject.status = "ERROR";
-//         returnObject.value = error;
-//         return chrome.tabs.sendMessage(sender.tab.id, returnObject);
-//     }
-
-// }
-async function sendPremiumStatus(steamId, sender, returnObject) {
-    try {
-        const resp = await fetch("https://rust-api.facepunch.com/api/premium/verify", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json"
-            },
-            body: JSON.stringify({ SteamIds: [steamId] }),
-        });
-        if (resp?.status !== 200) throw new Error(`Requesting premium status failed | steamId: ${steamId} | Status: ${resp?.status}`)
-
-        const data = await resp.json();
-        const value = { premium: data.Results[steamId] }
-
-        returnObject.status = "OK";
-        returnObject.value = value;
-        return chrome.tabs.sendMessage(sender.tab.id, returnObject);
-    } catch (error) {
-        console.error(error);
-        returnObject.status = "ERROR";
-        returnObject.message = error.message;
-        returnObject.value = error;
-        return chrome.tabs.sendMessage(sender.tab.id, returnObject);
-    }
-}
 async function sendProxyCheck(ips, apiKey, sender, returnObject) {
     try {
         const resp = await fetch(`https://proxycheck.io/v3/${ips}?key=${apiKey}`);
@@ -187,22 +145,7 @@ async function sendSteamPlayerBanSummaries(steamIds, API_KEY, sender, returnObje
         return chrome.tabs.sendMessage(sender.tab.id, returnObject);
     }
 }
-// async function sendPublicBans(steamId, apiKey, sender, returnObject) {
-//     try {
-//         const resp = await fetch(`https://rust-api.flqyd.dev/bans/${steamId}?accessToken=${apiKey}`);
-//         if (resp?.status !== 200) throw new Error(`Requesting Public Bans Failed | steamId: ${steamId} | API KEY: ${apiKey.substring(0, 10)}... | Status: ${resp?.status}`)
 
-//         const data = await resp.json();
-//         returnObject.status = "OK";
-//         returnObject.value = data.data.bans;
-//         return chrome.tabs.sendMessage(sender.tab.id, returnObject);
-//     } catch (error) {
-//         console.error(error);
-//         returnObject.status = "ERROR";
-//         returnObject.value = error;
-//         return chrome.tabs.sendMessage(sender.tab.id, returnObject);
-//     }
-// }
 async function sendAtlasTeaminfo(values, apiKey, sender, returnObject) {
     try {        
         const steamId = values.split("-")[0];