+++ bg/other/convertDate.js
@@ -1,5 +1,7 @@
 import { getRelativeTime } from "./getRelativeTime.js";
 
 export function convertDate(date) {
-    return `${date.toLocaleDateString()} (${getRelativeTime(date)})`;
+    const localDate = date.toLocaleDateString();
+    const relative = getRelativeTime(date);
+    return `${localDate} (${relative})`;
 }