+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [2.2.17]
+### Changed
+- Steam Hours -> All Games
+
+### Removed
+- Checking for premium accounts
+
 
 ## [2.2.16]
 ### Added